--- app/constants/constants.js
@@ -1,1 +1,0 @@
-export const urls = { MASTER_DATA: "masterData/all" };
